Excluded test generation for WasmPlugins examples (#15855)
[upstream:32aabcc667cbdeafc62242b261d3ccf5e09f757f]
