--- google/services/networkservices/resource_network_services_wasm_plugin_generated_test.go
@@ -1,206 +1,0 @@
-// Copyright (c) HashiCorp, Inc.
-// SPDX-License-Identifier: MPL-2.0
-
-// ----------------------------------------------------------------------------
-//
-//     ***     AUTO GENERATED CODE    ***    Type: MMv1     ***
-//
-// ----------------------------------------------------------------------------
-//
-//     This file is automatically generated by Magic Modules and manual
-//     changes will be clobbered when the file is regenerated.
-//
-//     Please read more about how to change this file in
-//     .github/CONTRIBUTING.md.
-//
-// ----------------------------------------------------------------------------
-
-package networkservices_test
-
-import (
-	"fmt"
-	"log"
-	"strconv"
-	"strings"
-	"testing"
-	"time"
-
-	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
-	"github.com/hashicorp/terraform-plugin-testing/terraform"
-
-	"github.com/hashicorp/terraform-provider-google/google/acctest"
-	"github.com/hashicorp/terraform-provider-google/google/envvar"
-	"github.com/hashicorp/terraform-provider-google/google/tpgresource"
-	transport_tpg "github.com/hashicorp/terraform-provider-google/google/transport"
-
-	"google.golang.org/api/googleapi"
-)
-
-var (
-	_ = fmt.Sprintf
-	_ = log.Print
-	_ = strconv.Atoi
-	_ = strings.Trim
-	_ = time.Now
-	_ = resource.TestMain
-	_ = terraform.NewState
-	_ = envvar.TestEnvVar
-	_ = tpgresource.SetLabels
-	_ = transport_tpg.Config{}
-	_ = googleapi.Error{}
-)
-
-func TestAccNetworkServicesWasmPlugin_wasmPluginBasicDockerRepositoryExample(t *testing.T) {
-	t.Skip("https://github.com/GoogleCloudPlatform/magic-modules/pull/12275#discussion_r2283175770")
-	t.Parallel()
-
-	context := map[string]interface{}{
-		"random_suffix": acctest.RandString(t, 10),
-	}
-
-	acctest.VcrTest(t, resource.TestCase{
-		PreCheck:                 func() { acctest.AccTestPreCheck(t) },
-		ProtoV5ProviderFactories: acctest.ProtoV5ProviderFactories(t),
-		CheckDestroy:             testAccCheckNetworkServicesWasmPluginDestroyProducer(t),
-		Steps: []resource.TestStep{
-			{
-				Config: testAccNetworkServicesWasmPlugin_wasmPluginBasicDockerRepositoryExample(context),
-			},
-			{
-				ResourceName:            "google_network_services_wasm_plugin.wasm_plugin",
-				ImportState:             true,
-				ImportStateVerify:       true,
-				ImportStateVerifyIgnore: []string{"labels", "location", "name", "terraform_labels"},
-			},
-		},
-	})
-}
-
-func testAccNetworkServicesWasmPlugin_wasmPluginBasicDockerRepositoryExample(context map[string]interface{}) string {
-	return acctest.Nprintf(`
-data "google_project" "project" {}
-
-resource "google_network_services_wasm_plugin" "wasm_plugin" {
-  name        = "tf-test-my-wasm-plugin%{random_suffix}"
-  description = "my wasm plugin from a docker repository"
-
-  main_version_id = "v1"
-
-  labels = {
-    test_label =  "test_value"
-  }
-  log_config {
-    enable =  true
-    sample_rate = 1
-    min_log_level =  "WARN"
-  }
-
-  versions {
-    version_name = "v1"
-    description = "v1 version of my wasm plugin"
-    image_uri = "us-central1-docker.pkg.dev/${data.google_project.project.name}/tf-test-my-artifact-registry-docker-repository%{random_suffix}/tf-test-my-wasm-plugin-data%{random_suffix}:prod"
-
-    labels = {
-      test_label =  "test_value"
-    }
-  }
-}
-`, context)
-}
-
-func TestAccNetworkServicesWasmPlugin_wasmPluginBasicGenericRepositoryExample(t *testing.T) {
-	t.Skip("https://github.com/GoogleCloudPlatform/magic-modules/pull/12275#discussion_r2283175770")
-	t.Parallel()
-
-	context := map[string]interface{}{
-		"random_suffix": acctest.RandString(t, 10),
-	}
-
-	acctest.VcrTest(t, resource.TestCase{
-		PreCheck:                 func() { acctest.AccTestPreCheck(t) },
-		ProtoV5ProviderFactories: acctest.ProtoV5ProviderFactories(t),
-		CheckDestroy:             testAccCheckNetworkServicesWasmPluginDestroyProducer(t),
-		Steps: []resource.TestStep{
-			{
-				Config: testAccNetworkServicesWasmPlugin_wasmPluginBasicGenericRepositoryExample(context),
-			},
-			{
-				ResourceName:            "google_network_services_wasm_plugin.wasm_plugin",
-				ImportState:             true,
-				ImportStateVerify:       true,
-				ImportStateVerifyIgnore: []string{"labels", "location", "name", "terraform_labels"},
-			},
-		},
-	})
-}
-
-func testAccNetworkServicesWasmPlugin_wasmPluginBasicGenericRepositoryExample(context map[string]interface{}) string {
-	return acctest.Nprintf(`
-data "google_project" "project" {}
-
-resource "google_network_services_wasm_plugin" "wasm_plugin" {
-  name        = "tf-test-my-wasm-plugin%{random_suffix}"
-  description = "my wasm plugin from a generic repository"
-
-  main_version_id = "v1"
-
-  labels = {
-    test_label =  "test_value"
-  }
-  log_config {
-    enable =  true
-    sample_rate = 1
-    min_log_level =  "WARN"
-  }
-
-  versions {
-    version_name = "v1"
-    description = "v1 version of my wasm plugin"
-    image_uri = "projects/${data.google_project.project.name}/locations/us-central1/repositories/tf-test-my-artifact-registry-generic-repository%{random_suffix}/genericArtifacts/tf-test-my-wasm-plugin-data%{random_suffix}:v1"
-
-    labels = {
-      test_label =  "test_value"
-    }
-  }
-}
-`, context)
-}
-
-func testAccCheckNetworkServicesWasmPluginDestroyProducer(t *testing.T) func(s *terraform.State) error {
-	return func(s *terraform.State) error {
-		for name, rs := range s.RootModule().Resources {
-			if rs.Type != "google_network_services_wasm_plugin" {
-				continue
-			}
-			if strings.HasPrefix(name, "data.") {
-				continue
-			}
-
-			config := acctest.GoogleProviderConfig(t)
-
-			url, err := tpgresource.ReplaceVarsForTest(config, rs, "{{NetworkServicesBasePath}}projects/{{project}}/locations/{{location}}/wasmPlugins/{{name}}")
-			if err != nil {
-				return err
-			}
-
-			billingProject := ""
-
-			if config.BillingProject != "" {
-				billingProject = config.BillingProject
-			}
-
-			_, err = transport_tpg.SendRequest(transport_tpg.SendRequestOptions{
-				Config:    config,
-				Method:    "GET",
-				Project:   billingProject,
-				RawURL:    url,
-				UserAgent: config.UserAgent,
-			})
-			if err == nil {
-				return fmt.Errorf("NetworkServicesWasmPlugin still exists at %s", url)
-			}
-		}
-
-		return nil
-	}
-}
